pyproject.toml
[project]
name = "features-agent"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.3.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-gigachat>=0.3.0",
    "gigachat>=0.1.40",
    "catboost>=1.2.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.5.0",
    "featuretools>=1.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "tabulate>=0.9.0",
]
